+++ app/src/main/java/com/jbrunoo/seoul_go/presentation/feature/home/HomeViewModel.kt
@@ -1,4 +1,4 @@
-package com.jbrunoo.seoul_go.presentation.feature.event
+package com.jbrunoo.seoul_go.presentation.feature.home
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -7,12 +7,11 @@ import com.jbrunoo.seoul_go.domain.useCase.GetEventUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-class EventViewModel @Inject constructor(private val getEventUseCase: GetEventUseCase): ViewModel() {
+class HomeViewModel @Inject constructor(private val getEventUseCase: GetEventUseCase): ViewModel() {
     private var _events = MutableStateFlow<List<Event>>(emptyList())
     val events = _events.asStateFlow()
